back/requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
beautifulsoup4
google-auth
python-jose[cryptography]